touch: fts: update to using gpiod APIs
Upstream is removing deprecated OF-specific gpio APIs in [1].
As a preparation, we need to update this driver here to use
the generic gpiod-based ones.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=17bbc46fc9d5

Note that in places where the value of the reset_gpio appears
to have changed this is not the case as the line is active low
but the old driver was ignoring that. The gpiod APIs will do
the right thing for us.

Test: touch works
Bug: 273738655
Change-Id: I5d563d31815c0345b6f5c7d42cb1ebd8f1d7d612
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Peter Griffin <gpeter@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>